Feature: Remove hard-coded tt_content table and CType
This patch moved all generation of TsConfig / TypoScript to events which
enabled us to check for TCA in the TableDefinitionCollection. By doing
that it is now possible to remove all hard-coded occurrences of
tt_content / CType. The type field is now optional, because tables can
also have just one default type / showitem "1".

Next steps are:
- Enable to generate custom type fields for custom tables.
- Automatically retrieve type of existing fields.

Change-Id: Iaf1b8ea72c40e6446f2f19fc4f6e8946973db038