Task: Use more sophisticated approach for adding typeField
The typeField is now added early on in the Factory. This has the
benefit, that SQL, TCA and showitems is built automatically. Some
extra checks are needed for the special typeField, though. Some
more tests are added along the way to ensure, the default value
is set for the first occurring record type.