[RELEASE] Release version 0.1.0
The first public TER release for Content Blocks! We've worked for 1 year
to get to this state. Please try it out and send feedback to us. This
will eventually become the official system extension content_blocks for
TYPO3 v13. As this is a beta version, be careful using it in production.
Things may break in every new release. Have fun using Content Blocks!