[TASK] Explicitly declare options to use array2xml
We should not use FlexFormTools here, because flexArray2Xml is used for
converting the dataStructure into DB storage format and many options are
not needed for our case. We only need the "el" to "field" mapping, so
that identifiers with dots are written into the "index" attribute of
"field". Else, dots are removed from identifiers.

Also add tests.