[FEATURE] Identifier replacement for displayCond
The identifiers for TCA displayCond are now replaced
with unique identifiers in case prefixing is enabled.
For this to work, a map of simple to unique identifiers
is collected during the main loop of fields. After the
main loop, a second round is started to check for
possible displayCond configs. If found, the config
is parsed (recursively for arrays syntax).