[FEATURE] Add first basic support for nested content elements
We integrate the basics for a default TYPO3 in order to prevent
rendering of nested elements in Page module within backend, as well as
styles.content.get TypoScript rendering for frontend.

We also document the technical approach and how to use the API for
custom implementations and other rendering approaches, e.g. via
DatabaseQueryProcessor.

"Basic" because we do not add rendering of the childs as ready to use
columns for frontend or backend, this will be part of #68.

Resolves: #106