[BUGFIX] Lift hard requirement for labelField type
It makes no sense to force the type for labelField and then falling
back to the first field, if the type does not match. The users' choice
is absolute. The fallback does only happen, if the user didn't define
labelField at all. The first preferred field is taken, otherwise the
first field in the list. FormEngine can handle most of the types well.

Fixes: #112