[BUGFIX] Always provide raw record in _raw
The raw record which is provided from RelationResolver
was only set for the root record correctly. Deeper
nesting levels incorrectly received the resolved record.
This is fixed now by providing the raw record in the
resolved record as _raw. The ContentBlockDataResolver
will read this property and build the new ResolvedRelation
transfer object from it. After that this internal property
is unset.

This fix allows to use nested Content Elements properly
and avoids array to string conversion arrays when the
raw record wasn't actually raw.