[TASK] Decouple relation resolving and data object decoration
The class ContentBlockDataResolver had too many responsibilities. It
initiated the actual relation resolving and on top did the
transformation to the data object format. This has now been decoupled
in order to adhere to the single responsibility principle.

With this refactoring it is now possible to use these two classes
separately. For example if the data object is not needed, the
transformation step can be skipped.

The class ContentBlockDataResolver is renamed to
ContentBlockDataDecorator to better reflect what it actually does.