[BUGFIX] Avoid NotInstantiableTypeException for type Folder
Objects of class TYPO3\CMS\Core\Resource\File can't be exported
due to attached storage, which itself has EventDispatcher attached
which eventually leads to illegal Closures for EventListeners.
Right now, this happens for relations of type "folder".

In order to fix this, the File object needs somehow to be defined
as lazy. Another way could be to define a specific Model which
behaves as a replacement for the File object.

Yet another fix could be to avoid having eventDispatcher as a property
in ResourceStorage. This is used in many places, though.

Fixes: #172