[FEATURE] Enable control over header and footer backend preview (#295)
This patch adds the possibility to override the standard
header and footer of Content Element previews. Before it
was only possible to add a preview for the content area.

For this to work, a new layout "Preview" has been added,
which all previews should include. This layout renders
three sections: Header, Content and Footer. One may
implement any of these sections. Omitting a section
will result to a fallback to the default preview
rendering of TYPO3. It is possible to effectively
disable a preview section, by rendering an empty preview.

The old preview without a layout still works, but is
now deprecated and will log a deprecation level log
entry. Users are advised to migrate now.

Fixes: #275