Added note about identifier uniqueness for collection fields
It is important to note, that it should be avoided to have multiple content blocks, which all have a field of the type `collection` and which all share the same table and which all have the same identifier.

If different content blocks use the same identifier for collection fields that use the same `foreign_table` (e.g. `tt_content`), it occurs, that collection children get mixed up.