[FEATURE] Type Overrides
This feature is basically the Content Blocks style implementation
of the mighty TCA feature `overrideChildTca`. This options allows to
override anything in the child table. An interesting use case for this
is to override the `showitem` string. This is especially useful when
dealing with the sys_file_reference table, as it does not allow to
define alternative types, leaving this the only way to have custom
fields. The only other alternative would be to completely override the
file type (which is already possible with Content Blocks as well).

Technically, Content Blocks is already able to do most of the stuff,
that we need for this:
- Convert the fields array to the showitem string
- Create new database columns for the table

Normally, new base fields are only created for new Content Types. But
here we only override these types. So we use a little trick by doing
the usual processing of fields and removing the type definition at the
end. What's left is the table definition, which will later be merged
into the final combined definition.