[BUGFIX] Avoid exception for missing label field in TCA generation
Content Blocks now won't throw an exception anymore if the type field
is missing or invalid while compiling TCA (which is not recoverable in
the backend). TYPO3 already shows an error in the backend when opening
the record in FormEngine.

It would be even better to have such a check in
TableDefinitionCollectionFactory, but this is a valid intermediate step.

Fixes: #353