[TASK] Introduce cache for SimpleTcaSchema (second attempt)
SimpleTcaSchemaFactory does two jobs at the same time. It acts as a
factory for itself and provides the API for the schema. For proper
cache handling and hydration, the cache must be only built once during
BeforeTcaOverridesEvent, as this is the only place, where the correct
base TCA is passed. It must never receive the full, global TCA array.
After the cache has been built, all consecutive constructions of this
class must look for an existing cache. It must never be the case where
the class is constructed without cache, but initialize was not called.

First TYPO3 boot / uncached boot:

1. Construct on empty cache
2. Fill the cache in BeforeTcaOverridesEvent with "initialize".

Cached TYPO3 boot:
1. Construct on filled cache

Fixes: #386