[TASK] Lazy Load
This change brings the much-needed lazy loading to ContentBlockData.
While the Core RecordObject already has the ability to lazily load
relations, Content Blocks negated it by reading all properties
recursively to determine how to handle them.

Lazy load is very important for database structures, which are heavily
interconnected. Even though this is not an infinite loop, it can easily
eat up all resources, which some real-world examples proved. As Content
Blocks is mainly used for self-contained Content Elements, this issue
was not as critical to this date.

With this patch, the same approach is taken, alike to Core. All relation
properties (also possible null values) are now stored in Closures and
only executed once they are accessed in ContentBlockData. Usually this
happens in Fluid context. Effectively this is another Closure for the
Core Closure.

The same had to be done for "Grids". This is a Content Blocks special
for rendering child relations in frontend and backend. For this the
helper ContentBlockGridData was added, which acts as another container.